feat: implement inventory movement modal and update product management functionality
