+++ views/dashboard.html
@@ -387,6 +387,6 @@
             </div>
         </div>
     </div>  
-    <script src="../assets/js/views/dashboard.js" type="module"></script>
+    <script src="../assets/js/functions/dashboard.js" type="module"></script>
 </body>
 </html>
