+++ views/inventario.html
@@ -305,8 +305,7 @@
         </div>
     </div>
     <script src="../assets/js/main.js"></script>
-    <script type="module" src="../assets/js/productos.js"></script>
-    <script type="module" src="../assets/js/dialog.js"></script>
+    <script type="module" src="../assets/js/inventario.js"></script>
     <script src="../assets/js/session/login.js"></script>
 </body>
